mam2: split and bazelify components
