+++ mam/v2/BUILD
@@ -11,9 +11,9 @@ cc_library(
     deps = [
         ":defs",
         ":list",
-        ":ntru",
-        ":pb3",
-        ":prng",
+        "//mam/v2/ntru",
+        "//mam/v2/pb3",
+        "//mam/v2/prng",
     ],
 )
 
@@ -24,7 +24,6 @@ cc_library(
         "//conditions:default": TRINARY_INT_1_TRITS_PER_WORD_DEFS,
     }),
     deps = [
-        "//common/trinary:trit_array",
         "//common/trinary:trit_long",
         "//common/trinary:trits",
     ],
@@ -37,116 +36,6 @@ cc_library(
 )
 
 cc_library(
-    name = "buffers",
-    srcs = ["buffers.c"],
-    hdrs = ["buffers.h"],
-    deps = [
-        "trits",
-        ":defs",
-        "//common/trinary:trit_array",
-    ],
-)
-
-cc_library(
-    name = "trits",
-    srcs = ["trits.c"],
-    hdrs = ["trits.h"],
-    deps = [
-        ":defs",
-        "//common/trinary:flex_trit",
-        "//common/trinary:trit_tryte",
-    ],
-)
-
-cc_library(
-    name = "prng",
-    srcs = ["prng.c"],
-    hdrs = ["prng.h"],
-    deps = [
-        ":defs",
-        ":err",
-        ":sponge",
-        ":trits",
-    ],
-)
-
-cc_library(
-    name = "sponge",
-    srcs = ["sponge.c"],
-    hdrs = ["sponge.h"],
-    deps = [
-        ":buffers",
-        ":defs",
-        ":trits",
-        "//common/trinary:add",
-    ],
-)
-
-cc_library(
-    name = "spongos",
-    srcs = ["spongos.c"],
-    hdrs = ["spongos.h"],
-    deps = [
-        ":sponge",
-    ],
-)
-
-cc_library(
-    name = "wots",
-    srcs = ["wots.c"],
-    hdrs = ["wots.h"],
-    deps = [
-        ":prng",
-        ":sponge",
-        ":spongos",
-    ],
-)
-
-cc_library(
-    name = "mss",
-    srcs = ["mss.c"],
-    hdrs = ["mss.h"],
-    deps = [
-        ":wots",
-        "//common:errors",
-    ],
-)
-
-cc_library(
-    name = "ntru",
-    srcs = ["ntru.c"],
-    hdrs = ["ntru.h"],
-    deps = [
-        ":poly",
-        ":prng",
-        ":spongos",
-    ],
-)
-
-cc_library(
-    name = "pb3",
-    srcs = ["pb3.c"],
-    hdrs = ["pb3.h"],
-    deps = [
-        ":mss",
-        "//common:errors",
-    ],
-)
-
-cc_library(
     name = "err",
     hdrs = ["err.h"],
-)
-
-cc_library(
-    name = "poly",
-    srcs = ["poly.c"],
-    hdrs = [
-        "poly.h",
-        "poly_param.h",
-    ],
-    deps = [
-        ":defs",
-        ":trits",
-    ],
 )
