+++ mam/v2/mss/BUILD
@@ -1,0 +1,10 @@
+package(default_visibility = ["//visibility:public"])
+
+cc_library(
+    name = "mss",
+    srcs = ["mss.c"],
+    hdrs = ["mss.h"],
+    deps = [
+        "//mam/v2/wots",
+    ],
+)
