+++ mam/v2/ntru/ntru.h
@@ -16,9 +16,9 @@
 #pragma once
 
 #include "mam/v2/defs.h"
-#include "mam/v2/prng.h"
-#include "mam/v2/spongos.h"
-#include "mam/v2/trits.h"
+#include "mam/v2/prng/prng.h"
+#include "mam/v2/sponge/spongos.h"
+#include "mam/v2/trits/trits.h"
 
 /*! \brief NTRU public key - 3g(x)/(1+3f(x)) - size. */
 #define MAM2_NTRU_PK_SIZE 9216
