+++ mam/v2/pb3/BUILD
@@ -1,0 +1,10 @@
+package(default_visibility = ["//visibility:public"])
+
+cc_library(
+    name = "pb3",
+    srcs = ["pb3.c"],
+    hdrs = ["pb3.h"],
+    deps = [
+        "//mam/v2/mss",
+    ],
+)
