+++ mam/v2/prng/BUILD
@@ -1,0 +1,13 @@
+package(default_visibility = ["//visibility:public"])
+
+cc_library(
+    name = "prng",
+    srcs = ["prng.c"],
+    hdrs = ["prng.h"],
+    deps = [
+        "//mam/v2:defs",
+        "//mam/v2:err",
+        "//mam/v2/sponge",
+        "//mam/v2/trits",
+    ],
+)
