+++ mam/v2/trits/trits.h
