+++ mam/v2/wots/BUILD
@@ -1,0 +1,12 @@
+package(default_visibility = ["//visibility:public"])
+
+cc_library(
+    name = "wots",
+    srcs = ["wots.c"],
+    hdrs = ["wots.h"],
+    deps = [
+        "//mam/v2/prng",
+        "//mam/v2/sponge",
+        "//mam/v2/sponge:spongos",
+    ],
+)
